+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## FIWARE Big Bang v0.6.0-next
 
+-   Add pre-install widgets (#83)
+
 ## FIWARE Big Bang v0.6.0 - 26 Oct, 2021
 
 -   Update Orion to version 3.3.0 (#80)